# Conflicts:
#	siga/src/main/webapp/WEB-INF/page/dpLotacao/lista.jsp

+++ siga-cp/src/main/java/br/gov/jfrj/siga/dp/AbstractCpOrgaoUsuario.java
@@ -58,7 +58,7 @@ public abstract class AbstractCpOrgaoUsuario extends Objeto implements Serializa
 	private Long idOrgaoUsu;
 
 	@Column(name = "CGC_ORGAO_USU")
-	private Integer cgcOrgaoUsu;
+	private Long cgcOrgaoUsu;
 
 	@Column(name = "COD_ORGAO_USU")
 	private Long codOrgaoUsu;
@@ -138,7 +138,7 @@ public abstract class AbstractCpOrgaoUsuario extends Objeto implements Serializa
 		return cepOrgaoUsu;
 	}
 
-	public Integer getCgcOrgaoUsu() {
+	public Long getCgcOrgaoUsu() {
 		return cgcOrgaoUsu;
 	}
 
@@ -208,7 +208,7 @@ public abstract class AbstractCpOrgaoUsuario extends Objeto implements Serializa
 		this.cepOrgaoUsu = cepOrgaoUsu;
 	}
 
-	public void setCgcOrgaoUsu(Integer cgcOrgaoUsu) {
+	public void setCgcOrgaoUsu(Long cgcOrgaoUsu) {
 		this.cgcOrgaoUsu = cgcOrgaoUsu;
 	}
